+++ src/test/java/test/TestSDTParser.java
@@ -54,21 +54,19 @@ public final class TestSDTParser {
 		s.s("S11", "transform { choose { when \"1\" { } when \"0\" { print \"0\" } } }", null);
 		s.s("S12", "transform { choose { when \"1\" { print \"1\" } otherwise { } } }", null);
 		s.s("S13", "transform { choose { when \"1\" { print \"1\" } otherwise { print \"0\" } } }", null);
-		s.s("S14", "transform { node \"a\" }", null);
-		s.s("S15", "transform { node \"a\" { } }", "transform { node \"a\" }");
-		s.s("S16", "transform { node \"a\" { node \"b\" } }", null);
-		s.s("S17", "transform { node \"a\" { value \"'b'\" } }", null);
-		s.s("S18", "transform { node \"a\" { println \"'b'\" } }", null);
-		s.s("S19", "transform { node \"a\" { value \"'b'\" node \"c\" copy \"'d'\" } }", null);
-		s.s("S20", "transform { copy \"/item\" }", null);
-		s.s("S21", "transform { foreach \"/i\" { sort \".\" } }", null);
-		s.s("S22", "transform { foreach \"/i\" { sort \".\" sort \".\" } }", null);
-		s.s("S23", "transform { foreach \"/i\" { sort \".\" print \".\" } }", null);
-		s.s("S24", "transform { foreach \"/i\" { sort \".\" sort \".\" print \".\" } }", null);
-		s.s("S25", "transform { foreach \"/i\" { sort \".\" { } print \".\" } }", "transform { foreach \"/i\" { sort \".\" print \".\" } }");
-		s.s("S26", "transform { foreach \"/i\" { sort \".\" { reverse \"1\" } } }", null);
-		s.s("S27", "transform { foreach \"/i\" { sort \".\" { comparator \"f(?,?)\" } } }", null);
-		s.s("S28", "transform { foreach \"/i\" { sort \".\" { comparator \"f(?,?)\" reverse \"0\" } } }", "transform { foreach \"/i\" { sort \".\" { reverse \"0\" comparator \"f(?,?)\" } } }");
+		s.s("S14", "transform { node \"a\" { } }", null);
+		s.s("S15", "transform { node \"a\" { value \"'b'\" } }", null);
+		s.s("S16", "transform { node \"a\" { println \"'b'\" } }", null);
+		s.s("S17", "transform { node \"a\" { value \"'b'\" node \"c\" { } copy \"'d'\" } }", null);
+		s.s("S18", "transform { copy \"/item\" }", null);
+		s.s("S19", "transform { foreach \"/i\" { sort \".\" } }", null);
+		s.s("S20", "transform { foreach \"/i\" { sort \".\" sort \".\" } }", null);
+		s.s("S21", "transform { foreach \"/i\" { sort \".\" print \".\" } }", null);
+		s.s("S22", "transform { foreach \"/i\" { sort \".\" sort \".\" print \".\" } }", null);
+		s.s("S23", "transform { foreach \"/i\" { sort \".\" { } print \".\" } }", "transform { foreach \"/i\" { sort \".\" print \".\" } }");
+		s.s("S24", "transform { foreach \"/i\" { sort \".\" { reverse \"1\" } } }", null);
+		s.s("S25", "transform { foreach \"/i\" { sort \".\" { comparator \"f(?,?)\" } } }", null);
+		s.s("S26", "transform { foreach \"/i\" { sort \".\" { comparator \"f(?,?)\" reverse \"0\" } } }", "transform { foreach \"/i\" { sort \".\" { reverse \"0\" comparator \"f(?,?)\" } } }");
 		
 		System.out.print("\n            "); /* test invalid SDT */
 		f.s("F01", "transfrom \"\"", "/transfrom: 'transform' statement expected");
@@ -133,12 +131,12 @@ public final class TestSDTParser {
 		f.s("F63", "transform { choose { when { print \"''\" } } }", "/transform/choose/when: statement 'when' requires an XPath expression");
 		f.s("F64", "transform { choose { when \"true()\" { print \"''\" } otherwise { print \"''\" }  when { } } }", "/transform/choose/otherwise: statement 'otherwise' is misplaced");
 		
-		f.s("F70", "transform { node \"\" }", "/transform/node: statement 'node' requires a node name");
+		f.s("F70", "transform { node \"\" }", "/transform/node: statement 'node' requires a compound statement");
 		f.s("F71", "transform { node { } }", "/transform/node: statement 'node' requires a node name");
 		f.s("F72", "transform { node \"a\" { select \"\" } }", "/transform/node/select: statement 'select' is not allowed here");
 		f.s("F73", "transform { node \"a\" { value \"\" } }", "/transform/node/value: statement 'value' requires an XPath expression");
 		f.s("F74", "transform { node \"2\" { value \"''\" } }", "/transform/node: node name '2' is invalid");
-
+		
 		f.s("F80", "transform { copy \"\" }", "/transform/copy: statement 'copy' requires an XPath expression");
 		f.s("F81", "transform { copy { } }", "/transform/copy: statement 'copy' expects no compound statement");
 		f.s("F82", "transform { copy \"/item\" { } }", "/transform/copy: statement 'copy' expects no compound statement");